add simpleConsistency and getMeta methods to test apis
