+++ package.json
@@ -16,7 +16,7 @@
     "prepublishOnly": "npm test",
     "test": "ts-node test | tap-diff",
     "test:unit": "ts-node test/unit | tap-diff",
-    "test:e2e": "ts-node test/e2e"
+    "test:e2e": "ts-node test/e2e | node node_modules/faucet/bin/cmd.js"
   },
   "author": "",
   "license": "ISC",
@@ -42,6 +42,7 @@
     "@types/lodash": "^4.14.144",
     "@types/node": "^12.7.5",
     "@types/ramda": "^0.26.33",
+    "faucet": "0.0.1",
     "sinon": "^7.4.2",
     "tap-diff": "^0.1.1",
     "tape": "^4.11.0",
